services/requirements.txt
Flask==2.1.0
gunicorn==19.3.0
requests==2.25.0
Pillow==9.1.0
tensorflow==2.9.1